+++ src/panopticas/cli.py
@@ -1,11 +1,12 @@
-# panopticas.py
+# panopticas CLI
 import click
 from prettytable import PrettyTable
-import panopticas as ft
+from . import core
+from .constants import VERSION
 
 
 @click.group()
-@click.version_option(version=ft.VERSION)
+@click.version_option(version=VERSION)
 def cli():
     """Panopticas is a tool for identifying file types, code and git repositories.
 
@@ -29,7 +30,7 @@ def assess(directory,unknown):
     else:
         click.echo('Assessing current directory.')
         directory = "."
-    files = ft.identify_files(directory)
+    files = core.identify_files(directory)
     click.echo(f'Found {len(files)} files.\n')
     table = PrettyTable()
     table.field_names = ["File", "Language", "Meta"]
@@ -38,7 +39,7 @@ def assess(directory,unknown):
     table.align["Meta"] = "l"
 
     for file, file_type in files.items():
-        meta = ft.get_filename_metatypes(file) if ft.get_filename_metatypes(file) else ""
+        meta = core.get_filename_metatypes(file) if core.get_filename_metatypes(file) else ""
         if meta:
             meta = ", ".join(meta)
         # with "unknown", we only include files with unknown language types (e.g. None)
@@ -63,16 +64,16 @@ def identify(file):
     table.align["Method"] = "l"
     table.align["Result"] = "l"
 
-    table.add_row(["File extenion", ft.get_fileext(file)])
-    table.add_row(["File type", ft.get_extension_filetype(ft.get_fileext(file))])
-    shebang = ft.check_shebang(file)
+    table.add_row(["File extenion", core.get_fileext(file)])
+    table.add_row(["File type", core.get_extension_filetype(core.get_fileext(file))])
+    shebang = core.check_shebang(file)
     table.add_row(["Shebang", shebang])
-    table.add_row(["Shebang Language", ft.extract_shebang_language(shebang) if shebang else None])
-    table.add_row(["Meta", ft.get_filename_metatypes(file)])
+    table.add_row(["Shebang Language", core.extract_shebang_language(shebang) if shebang else None])
+    table.add_row(["Meta", core.get_filename_metatypes(file)])
 
     urls = []
 
-    urls = ft.extract_urls_from_file(file)
+    urls = core.extract_urls_from_file(file)
     table.add_row(["URLs", '\n'.join(urls)])
 
     print(table)
@@ -85,16 +86,16 @@ def find_urls(directory,all_files):
     """
     Find and show urls for all files in a given directory.
     """
-    files = ft.find_files(directory,all_files=all_files)
+    files = core.find_files(directory,all_files=all_files)
 
     table = PrettyTable()
     table.field_names = ["Filename", "URLs"]
     table.align["Filename"] = "l"
     table.align["URLs"] = "l"
 
-    # ft.find_files(directory)
+    # core.find_files(directory)
     for f in files:
-        urls = ft.extract_urls_from_file(f)
+        urls = core.extract_urls_from_file(f)
         table.add_row([f, '\n'.join(urls)])
 
     print(table)